feat(ui): merge settings into a single profile page
Fold preferences into Profile, drop the Settings screen, and move the
profile/sign-in control to the bottom of the sidebar.
